## Overview
### What is the feature?
This change involved updating a few files, centered mainly around the design/layout of the app, a proper image asset for the launcher and an update to the README.md to include some screenshots of the app in action.

### What is the planned solution?
The solution to update the app was carried out by refactoring the above mentioned files.